+++ Magic/CHANGELOG.md
@@ -31,6 +31,7 @@
  - Add `Respawn` action, for respawning mobs that have died
  - Add `testdummy` mob, a loveable punching bag you can use for spell testing
  - Add support for getting colored tipped arrows, `/mgive tipped_arrow{18001}:#00FF00`
+ - Add `leash_range` mob property, currently only works for NPCs
 
 ## Changes